+++ src/Curiosity.Tools.Web/SiteMap/SiteMapExtraPageProvider.cs
@@ -11,7 +11,7 @@ namespace Curiosity.Tools.Web.SiteMap
         /// <summary>
         /// Provides extra pages for site map.
         /// </summary>
-        public virtual ICollection<(string url, DateTime lastChangeData)> GetExtraPages()
+        public virtual IReadOnlyList<(string url, DateTime lastChangeData)> GetExtraPages()
         {
             return Array.Empty<(string, DateTime)>();
         }